requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
cloudinary==1.41.0 